+++ example/filetransfer/Download3.cpp
@@ -37,7 +37,7 @@ void onConnection(const TcpConnectionPtr& connection)
             connection->setContext(fptr);
             char buf[kBuffSize];
             size_t nread = fread(buf, 1, sizeof(buf), fp);
-            connection->send(buf, nread);
+            connection->send(buf, static_cast<int>(nread));
         }
         else
         {
@@ -54,7 +54,7 @@ void onWriteComplete(const TcpConnectionPtr& connection)
     size_t nread = fread(buf, 1, sizeof(buf), boost::get_pointer(fptr));
     if (nread > 0)
     {
-        connection->send(buf, nread);
+        connection->send(buf, static_cast<int>(nread));
     }
     else
     {
